Make backend API URL configurable via .env file
The value in .env for SNOWPACK_PUBLIC_BACKEND_PROD_API_URL will be
used when the snowpack deployment mode is 'production' which it will
be when 'snowpack build' is invoked via 'npm run build' (due to config
in package.json). Otherwise a value for API URL is used which is
compatible with local development, i.e., when snowpack deployment mode
is 'development' which it is when 'snowpack dev' is invoked via 'npm
start' (due to config in package.json).