Move to monorepo with frontend changes too
* Change to monorepo

We are no longer using a separate repo for the frontend and
associated submodule. All commits have been replayed onto the monorepo
to preserve everyone's work and additional changes have been made
related to upgrades.

* Change from snowpack to vite

The svelte community has abandoned snowpack in favor of the vite build
and runtime system. This change improved the system both in terms of
build speed and also solved a problem wherein use of environment
variables in snowpack was broken.

* Frontend now uses a string table in the form of .env and .env.production

Vite provides the ability to inject environment variables defined in
.env and .env.production (See
https://vitejs.dev/guide/env-and-mode.html#env-variables).
./frontend/src/App.svelte has been updated to use this feature. Now
all text that is displayed in the user interface can be changed by
programmers or non-programmers by changing the values in .env and
.env.production and then doing a 'docker-compose build' or
equivalently a 'make build'.