Add bible commentary feature
* This commit adds English bible commentary feature (1st version). For
now the linking to bible commentary articles is transformed to point
to the article resources located at
https://content.bibletranslationtools.org/WycliffeAssociates/en_bc/src/branch/master/{}

Pending review by stakeholders of this linking strategy we may choose
to integrate said linked articles to the end of the PDF and HTML
documents as per translation words. For now the main commentary is
rendered, but, as said above, the links to articles point to those
resources online.