Use ProcessPoolExecutor where it makes sense
The advice normally given is that IO bound is well served by
threads while CPU is well served by processes. In practice for this
project, I have found that Python's ProcessPoolExecutor gives the best
results of the two for all tasks. Why? Probably because spawning
child processes in Python is the way you are going to fully utilize
all cores on the CPU and simultaneously avoid the global interpreter
lock (GIL). Having tried both and a combination of both in the ways
normally advised (as above) found that indeed ProcessPoolExecutor is
always faster for our cases here.