Update package dependencies
Do some house cleaning unrelated to issue with broken build. Just
removing some dependencies that are no longer used and updating bs4 to
beautifulsoup4 which is what it should be using. The broken build was
not experienced locally in Docker build, couldn't replicate, but
regardless a "fix" was introduced in tag 0.0.8,
https://github.com/linearcombination/USFM-Tools/commit/dcdcf9f48d0d14d2763bc5eaf0c798bbf6803c19
that relaxes version numbers so that pip can resolve to the versions
needed in DOC.