Run e2e tests in Docker in parallel
This used to be the only place we didn't run pytest in parallel
because there was an undiagnosed issue with doing so (although it was
fine outside Docker, go figure). Now that Celery
is being used and the execution model has changed this issue seems to
have evaporated, so we'll take the speedup in test suite runs.