Lengthen playwright test timeouts
To allow tests to complete. Creating large PDFs consisting of multiple
languages and possibly multiple books from start to finish is often a
multi-minute experience due to an elongation in the following stages:
asset location lookup, asset provision, asset parsing, asset assembly,
and conversion to final output format. The first two tasks are subject
o network I/O, and the penultimate two are subject to CPU bound
activities, while the last is subject to wkhtmltopdf binary conversion
which is done with a headless instance of webkit and takes longer than
we'd prefer.