Mark certian tests with a pytest custom marker
Register and use the custom pytest marker 'randomized'. Tests that are
random in nature and that we expect could fail (due to their ability,
thropugh randomness, to discover combinations to test that have not
been tested before) aren't suitable to run in a ci/cd production
situation. We instead run them in development periodically as desired.
To do so requires that we exclude them when we want to and pytest
markers are the means to do that.