Take down Docker compose processes before building
Developer ergonomics to type less when using Make targets. It isn't
strictly required to stop Docker processes before a build as they'll
be recreated when a build cache miss occurs, but if you
change a .env file and don't do this perhaps no build will get triggered since
.env files don't get built and thus the Docker processes don't get
recreated on your next test run and thus you could subsequently be
testing old state which will not reflect changes to the .env file, so
let's just stop Docker processes with our 'down' Makefile target each
time we build to be on the safe side. I was bit by this.