Add weasyprint to convert certain layouts to PDF
wkhtmltopdf cannot handle the column-count directive which we use to
flow in two columns TN and TQ content for the 1c layout to reduce
whitespace. weasyprint is slower, but handles these layouts. However,
weasyprint doesn't have the same header and footer functionality as
wkhtmltopdf so it is possibly only a temporary solution until a
different, probably less clean, way of handling a dynamic amount of
text flowing into a two column layout that can traverse page
boundaries is discovered that wkhtmltopdf can handle.