Draft github action
Just a draft untested github action that uses the actual steps
necessary to build and test both the API and the UI that work
locally (but may need adjusting in the github action context). This
should be useful to cherry pick, not in the git sense, from to build
the next version of the github action. It doesn't use qemu or buildx
steps or anything like that, it is just for illustrative purposes. buildx
may not be needed if we don't need to build for multiple platforms,
but that is just a side note.