Ensure valid state of output format switches
When you select a non-print layout, say ePub, and then subsequently,
in the same UI session after the document request is complete, choose
a print format, the UI will correctly hide the ePub and Docx output
format choices. But just because you now cannot see them does not mean
that there state is set to a valid value (only PDF) for print
optimization format. I.e., make sure if print optimization is chosen
that ePub and Docx are set to false.