Add initial proper docx support
Docx is the new preferred document output format as it is more easily
editable for translators.

This commit adds native docx support and an associated path for it in
the UI and backend as well as a test suite for it.

Supporting Docx meant adding a new code path that doesn't convert from
a finished HTML document to Docx in a wholesale fashion (in comparison
with conversion to PDF or ePub). Instead, the Docx is built up from
HTML snippets intermingled with required low-level Docx manipulations
to control one or two column layout and ultimately Docx template use
too. In turn this means that certain UI changes were necessary. The UI
changes include making the user choose either PDF, ePub, or Docx
rather than being able to choose all of them at once if desired. This
has the added benefit that a user can no longer choose to create three
different output formats in one go and thus the time for the request
to be fulfilled will be greatly reduced in such cases which will also
make the user feel the system is snappier (objectively, but more
importantly, subjectively).

Another UI change not directly related to Docx support is that we have
removed the drop down control to select chunking size. Instead, the UI
always sends chapter chunk size now and never verse chunking. Verse
chunking was decided to be of no value for translation purposes at
this time and verse interleaving is very costly in terms of
performance which makes for a bad UX. It can be resuscitated from git
history later if we want it back in the future though it is hard to
imagine now that happening.

Note too that Docx support does not include one of the layout options
available to HTML, PDF, and ePub: two column side by side languages.
This was never vetted as a needed requirement and has not been used or
requested for some time. It still exists as an option in HTML, PDF,
and ePub output formats as sort of a showcase for what can be done.
Also this layout would be much more difficult to implement in Docx.

This commit does not yet accommodate fonts for certain languages in
Docx format. To do so in Word will have to be the subject of a future
PR and appears to be of some complication. More research is needed.

There are also a few miscellaneous commits related to better naming,
a fix to configuration in docker-compose-override.yml, changing chunking
of several tests from the no longer used verse chunk size to chapter
chunk size instead, and remove commented out link in a Jinja HTML
template.

In a coming PR we will likely also want to ensure sources present no
linking in the generated doc. For now, for some sources, there are
some dead links in PDF, ePub, and Docx output formats. There is a
great deal of heterogeneity in links in the source. We successfully
handle tens of types, but there are apparently yet more that exist and
sometimes they are just malformed in the sources.