Set default for assembly strategy rather than give choice
Previous to this commit the UI allowed a user to choose to 'separate'
or 'mix' content. Now content is set to always separate.

This means you can still request multiple (a max of two) languages in
one document, but you'll see all the content for one language and then
all the content for the other language rather than mixing one
chapter (scripture and helps) of one language and then a chapter of
another language.

The 'mix' strategy could be brought back later, but mostly existed to
allow the user to choose the scripture left, helps right layout
strategy which has since been put aside as a non-requirement.

The net effect of this commit then is to also simplify the UI yet
again for the full version of the UI so that the user has one less
choice to make and yet the documents produced still satisfy all
requirements we have heard expressed thus far.