Don't enable scripture left, helps right layout at this time
With scripture left, helps right being a non-requirement for this
release we've disabled it. We learned today that it might likely be a
future requirement. Because a visual defect in the layout was
discovered recently we'll need to do some work on it and enable it
again for a future release if it does indeed become a required layout
option.