Config sveltekit to use static adaptor
We don't need node as a proxy anymore. We are using sveltekit in a spa
config. This means that the nginx config changes too and likely the
issue we were having with the nginx uri having a trailing slash
appended before URL params (which caused a redirect and error in
runtime HTTPS browser checked validation) goes away.