Fix redirect issue
All the URLs in translations.json were changed to use the transfer
URLs (from BIEL to DOC). So all requests travel through the frontend's
transfer route. The transfer route pulls apart the passed URL and
gathers assets based on the URL. When it came to getting resource
types, it was (unlike the /resource_types route) adding a trailing
slash just prior to the query string portion of the URL.