Fix possible bug
Bug hasn't been reported or experienced, but the logic should only set
the resource type to 'ulb-wa' when the second component is 'ulb' AND
the language code is 'en' otherwise all transfer URL's of the form
[lang_code]_ulb would have their resource set to 'ulb-wa' which would
obviously be incorrect. I think I meant to do that and had just
forgotten to add the check for language code being equal to 'en'.