Better handling of 'bc' resource type content
BC content often had seemingly erroneous or extraneous/accidental
pagination-ish symbols, | >>. Remove them.

BC content has links to articles that are meant to be auxiliary
material. In the future we may support such links, but for now we
deactivate them. In the 'read on web' version on BIEL they are shown
in popup window with mouseover.