Improve selection of resource types Prior to this commit, selection of resource types could result in a situation where the user could be presented with a resource type that is not actually available for the books they have chosen. For example, the user may have chosen English and an OT book and a BC resource type among others. In such a case, the resulting document would have still been generated, but the BC resource would be missing from the document since the BC resource does not provide assets for OT books, only NT.
With this commit, the resource type list presented to the user is a
product of deeper interrogation into the actual assets to see if the
resource covers the books they have chosen in the previous step. This
means, effectively, that the resource types step in the wizard takes a
little longer if the resource assets have not yet been acquired, but
also means that the subsequent step, when the document request is
submitted, will be much faster because the resource asset acquisition
has already occurred. Messaging to the user on the resource types
progress widget has been updated too to make the user aware.