Handle NULL bytes and control characters in source USFM
Our test team noted that when building a doc for English->Lingala (ln), the build fails with an error

This happened because the USFM for Lingala had control characters in
the text, in particular ^B (STX in ASCII - start of text). Since there
could be other source languages that have control characters I have
augmented STET to strip them out prior to processing.