Fix repeated verse issue Translation Services noted that when rendering the document for English -> Kubu (kvb), some target text verses are repeated.
This was the classic "in a loop not using old value since current
value is null". Fixed, as always, with freshly initializing value with
empty string so that the value established in the last loop iteration
is not used. Why were some verses null? Because the target language,
in this case, did not have a particular book or verse available.