Handle different types of USFM layouts on disk differently
USFM layout in repo comes in mostly two forms:

- one file per book, e.g., 1-GEN.usfm, or
- one directory per chapter, with one file per verse span in that
  directory (which we pull together into one file per book)

The latter layout projects a little more localization using
<root>/front/title.txt and <root>/<chapter>/title.txt files which in
turn results in often different (more complete) use of USFM chapter
labels. Because of this, the two types have to be handled differently
when splitting into chapters as the regex that is splitting hinges on
is different for each case.