Remove header from docx template
Because docx layout is configurable, there is currently no consistent
docx styled element that a template header could count on being there.
This would mean that when such an element is not there, you get an
error in the header instead. Eventually maybe we will have dynamic
header logic for docx.