Add a couple log outputs
Help to see when there is a potentially a problem with a STET input
file by examining the book codes it supposedly contains in its
references. If the book codes don't include what you expect it could
be that the book name in the STET input doc does not match the
expected localized name in DOC. E.g., DOC (actually translators) use
Filemom but the pt-br STET input doc uses Filemon.