Re-enable cache, but only for 30 minutes
This is the cache for file artifacts which include graphql query
results, resources.json, git repos, and generated document files. 30
minutes is long enough to materially impact running of the test suite
to make it considerably faster and short enough to allow new updates
to data to become available without too long of a wait.