Ensure assets are in place after startup init
Made this change because github CI is failing in a way that seems to
indicate a missing en_rg_nt_survey.docx as a possibility even though
it succeeds locally. The reason this could be the case is that github
CI might not persist assets between runs and perhaps in that
environment the asset is not actually being copied into place.