Gracefully handle case where data API returns null resource_type
This is a questionable change because it means that we don't catch
invalid data in the data API as quickly for the case when
content.resource_type is null (which is an invalid state). However,
without this commit, if the data API returns content.resource_type
null ever again, it zombies this app and others that depend on it as
an API which is bad for users. After this commit this type of error
will only be detected by a DOC user if they were expecting to see the
offending record and it isn't present in the results.