New strategies, layouts, linking, and context aware options
This is a big commit. The primary deliverable for this commit is
adding versification at the book and chapter chunk level for PDF,
ePub, and Docx, their supporting front and backend tests, and new
layouts that match the document prototype provided by the PO as a
guide.

This commit also makes TW words linking work properly.

This commit also adds new optional settings that the user may select
which are contextually relevant for the given assembly
strategy (interleaving strategy chosen). These can be seen dynamically
changing on the final settings page of the UI depending on what is
chosen.

Many bug fixes and visual enhancements to layout along the way.

More to be done, but this should be a radically useful upgrade for TS.